[Producer] Slight modularization
